Added changes after conflict resolution
