+++ src/test/java/baseEntities/BaseTest.java
@@ -4,7 +4,6 @@ import browserFactory.BrowserFactory;
 import org.openqa.selenium.WebDriver;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
-import steps.AddingUserStep;
 import steps.AddUserStep;
 import steps.LoginStep;
 import steps.UsersStep;
@@ -14,7 +13,6 @@ public class BaseTest {
     protected WebDriver driver;
 
     protected LoginStep loginStep;
-    protected AddingUserStep addingUserStep;
     protected UsersStep usersStep;
     protected AddUserStep addUserStep;
 
@@ -26,7 +24,6 @@ public class BaseTest {
         usersStep = new UsersStep(driver);
         addUserStep = new AddUserStep(driver);
         driver.get(ReadProperties.getUrl());
-        addingUserStep = new AddingUserStep(driver);
     }
 
     @AfterMethod
