+++ src/test/java/steps/AddUserStep.java
@@ -5,6 +5,8 @@ import models.User;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.openqa.selenium.WebDriver;
+import pages.AddUserPage;
+import pages.SettingsPage;
 import pages.UsersPage;
 
 public class AddUserStep extends BaseStep {
@@ -20,5 +22,23 @@ public class AddUserStep extends BaseStep {
         addUserPage.addUser(user);
         logger.info("New user is added");
         return usersPage;
+    }
+
+    public SettingsPage addingProjectSettings() {
+        projectsPage.clickToSettings();
+
+        return settingsPage;
+    }
+
+    public UsersPage addingUser() {
+        addingProjectSettings().clickToUsersMenuLabel();
+
+        return usersPage;
+    }
+
+    public AddUserPage openUserForm() {
+        usersPage.clickAddUserButton();
+
+        return addUserPage;
     }
 }
